requirements.txt
google-generativeai
python-dotenv
selenium
Pillow
PyPDF2
pdf2image
streamlit